moved generator ring config to a sharead area
