+++ pkg/ring/config.go
@@ -1,4 +1,4 @@
-package generator
+package ring
 
 import (
 	"flag"
@@ -15,7 +15,7 @@ import (
 	"github.com/grafana/tempo/pkg/util/log"
 )
 
-type RingConfig struct {
+type Config struct {
 	KVStore          kv.Config     `yaml:"kvstore"`
 	HeartbeatPeriod  time.Duration `yaml:"heartbeat_period"`
 	HeartbeatTimeout time.Duration `yaml:"heartbeat_timeout"`
@@ -30,7 +30,7 @@ type RingConfig struct {
 	ListenPort int `yaml:"-"`
 }
 
-func (cfg *RingConfig) RegisterFlagsAndApplyDefaults(prefix string, f *flag.FlagSet) {
+func (cfg *Config) RegisterFlagsAndApplyDefaults(prefix string, f *flag.FlagSet) {
 	cfg.KVStore.RegisterFlagsWithPrefix(prefix, "collectors/", f)
 	cfg.KVStore.Store = "memberlist"
 
@@ -46,7 +46,7 @@ func (cfg *RingConfig) RegisterFlagsAndApplyDefaults(prefix string, f *flag.Flag
 	cfg.InstanceInterfaceNames = []string{"eth0", "en0"}
 }
 
-func (cfg *RingConfig) ToRingConfig() ring.Config {
+func (cfg *Config) ToRingConfig() ring.Config {
 	rc := ring.Config{}
 	flagext.DefaultValues(&rc)
 
@@ -58,7 +58,7 @@ func (cfg *RingConfig) ToRingConfig() ring.Config {
 	return rc
 }
 
-func (cfg *RingConfig) toLifecyclerConfig() (ring.BasicLifecyclerConfig, error) {
+func (cfg *Config) ToLifecyclerConfig(numTokens int) (ring.BasicLifecyclerConfig, error) {
 	instanceAddr, err := ring.GetInstanceAddr(cfg.InstanceAddr, cfg.InstanceInterfaceNames, log.Logger, cfg.EnableInet6)
 	if err != nil {
 		level.Error(log.Logger).Log("msg", "failed to get instance address", "err", err)
@@ -73,6 +73,6 @@ func (cfg *RingConfig) toLifecyclerConfig() (ring.BasicLifecyclerConfig, error) 
 		ID:              cfg.InstanceID,
 		Addr:            instanceAddrPort,
 		HeartbeatPeriod: cfg.HeartbeatPeriod,
-		NumTokens:       ringNumTokens,
+		NumTokens:       numTokens,
 	}, nil
 }
